Plot standardized mean differences

Quickly plot standardized mean differences. Plots are faceted by group.

# S3 method for smd
autoplot(object, method = "iptw_att")

# S3 method for smd_list
autoplot(object, method = "iptw_att", id = NULL)

Arguments

object
An object of the appropriate class.

method
A character vector of length one indicating the propensity score method used to generate the weights.

id
Character vector of the same length as object to name each element in the list of smd objects. Passed to the .id argument in dplyr::bind_rows().

Value

A ggplot2::ggplot object.

Details

If method is NULL, then all methods are compared in a single plot. However, in this case it is infeasible to display each covariate in the plot so only the (logit of the) propensity scores are compared. If object is of class smd_list, then SMDs are compared (via the colour aesthetic) across all objects in the list.