ecmeta.nsclc: Application of a Meta-Analytic Framework for Decision Making and Error Control to External Control Analyses in NSCLC

A proof of concept in non-small cell lung cancer (NSCLC) of a meta-analytic framework to adjust external control studies for additional bias and variability due to the non-randomized design. A total of 14 pairwise comparisons are reanalyzed using propensity score methods as the basis for application of the framework.

See also

Author

Maintainer: Devin Incerti incerti.devin@gene.com

Authors:

Michael Bretscher michael.bretscher@roche.com

Ray Lin lins24@gene.com

Chris Harbron chris.harbron@roche.com