HTML analysis table

Create an HTML table using using knitr::kable() (and styled with kableExtra::kable_styling()) that summarizes an analysis object. For each analysis, it indicates the external control data source, the clinicaltrials.gov identifier for the clinical trial, the experimental and comparator arm, and the sample size for both the external control and the experimental arm of the clinical trial.

# S3 method for analysis
html_table(x)