Specification of propensity score models.

A lookup table used to determine the specification of the propensity score model for each analysis.

psmodel_specification

Format

A data.frame. The first two columns are identifiers for an analysis. The remaining columns are potential covariates for the propensity score model: a variable is included if it has a value of 1 and is excluded if it has a value of 0.

See also

Examples

ecmeta.nsclc::psmodel_specification

#>    analysis_num      analysis_id sex race_grouped age smoker_grouped histology
#> 1             1 nct02008227_fi_1   1            1   1              1         1
#> 2             2 nct01903993_fi_1   1            1   1              1         1
#> 3             3 nct02366143_fi_1   1            1   1              1         0
#> 4             4 nct01351415_fi_1   1            1   1              0         0
#> 5             5 nct01519804_fi_1   1            0   1              1         0
#> 6             6 nct01496742_fi_1   1            1   1              1         0
#> 7             7 nct01496742_fi_2   1            1   1              1         0
#> 8             8 nct01366131_fi_1   1            1   1              0         0
#> 9             9 nct01493843_fi_1   1            1   1              0         0
#> 10           10 nct01493843_fi_2   1            1   1              0         0
#> 11           11 nct01493843_fi_3   1            1   1              0         0
#> 12           12 nct02367781_fi_1   1            1   1              1         0
#> 13           13 nct02367794_fi_1   1            1   1              1         0
#> 14           14 nct02657434_fi_1   1            1   1              1         0
#>    stage_grouped days_since_dx interaction
#> 1              1             1           1
#> 2              1             1           1
#> 3              1             1           1
#> 4              1             1           0
#> 5              1             1           0
#> 6              1             1           0
#> 7              1             1           0
#> 8              1             1           1
#> 9              0             1           0
#> 10             0             1           0
#> 11             0             1           0
#> 12             1             1           1
#> 13             1             1           1
#> 14             1             1           1